# server/openapi.yaml
openapi: 3.0.3
info:
  title: Risk Register API
  version: 1.0.0
  description: >-
    REST-API zur Verwaltung eines Risikoregisters.


    Die API unterstützt das Erfassen, Suchen, Aktualisieren und Löschen von
    Risiken.

    Risikowerte liegen zwischen 1 und 25 und ergeben sich fachlich aus
    Eintrittswahrscheinlichkeit × Auswirkung.
  contact:
    name: Risk Register Team
  license:
    name: Private use
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: signed-token
  schemas:
...
    ErrorResponse:
      title: ErrorResponse
      description: Einheitliche Fehlerantwort der Risk-Register-API.
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - RISK_NOT_FOUND
            - MEASURE_NOT_FOUND
            - USER_NOT_FOUND
            - AUTH_FAILED
            - FORBIDDEN
            - INTERNAL_ERROR
          example: RISK_NOT_FOUND
        message:
          type: string
          example: Das Risiko mit der angegebenen ID wurde nicht gefunden.
      required:
        - code
        - message
...
    User:
      title: User
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          description: Technische ID des Benutzers.
          example: 01J00000000000000000001000
        username:
          type: string
          minLength: 3
          maxLength: 80
          pattern: ^[a-zA-Z0-9._-]+$
          description: Eindeutiger Benutzername für die Anmeldung.
          example: admin
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Anzeigename des Benutzers.
          example: Admin User
        email:
          type: string
          format: email
          maxLength: 254
          description: E-Mail-Adresse des Benutzers.
          example: admin@example.local
        department:
          type: string
          minLength: 1
          maxLength: 200
          description: Abteilung des Benutzers.
          example: Risk Management
        role:
          type: string
          enum:
            - Admin
            - User
          description: Rolle des Benutzers. Nur Admins dürfen Einstellungen und Benutzer
            verwalten.
          example: Admin
        active:
          type: boolean
          description: Steuert, ob sich der Benutzer anmelden darf.
          example: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: 2026-07-01T08:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: 2026-07-01T08:00:00.000Z
      required:
        - id
        - username
        - name
        - email
        - department
        - role
        - active
        - createdAt
        - updatedAt
      example:
        id: 01J00000000000000000001000
        username: admin
        name: Admin User
        email: admin@example.local
        department: Risk Management
        role: Admin
        active: true
        createdAt: 2026-07-01T08:00:00.000Z
        updatedAt: 2026-07-01T08:00:00.000Z
    CreateUser:
      title: CreateUser
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 80
          pattern: ^[a-zA-Z0-9._-]+$
          description: Eindeutiger Benutzername für die Anmeldung.
          example: admin
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Anzeigename des Benutzers.
          example: Admin User
        email:
          type: string
          format: email
          maxLength: 254
          description: E-Mail-Adresse des Benutzers.
          example: admin@example.local
        department:
          type: string
          minLength: 1
          maxLength: 200
          description: Abteilung des Benutzers.
          example: Risk Management
        role:
          type: string
          enum:
            - Admin
            - User
          description: Rolle des Benutzers. Nur Admins dürfen Einstellungen und Benutzer
            verwalten.
          example: Admin
        active:
          type: boolean
          description: Steuert, ob sich der Benutzer anmelden darf.
          example: true
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: Passwort des Benutzers. Wird nie in API-Antworten zurückgegeben.
          example: admin123
      required:
        - username
        - name
        - email
        - department
        - password
      example:
        username: lena.vogt
        name: Lena Vogt
        email: lena.vogt@example.local
        department: Risk Management
        role: User
        password: sicher123
        active: true
    UpdateUser:
      title: UpdateUser
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 80
          pattern: ^[a-zA-Z0-9._-]+$
          description: Eindeutiger Benutzername für die Anmeldung.
          example: admin
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Anzeigename des Benutzers.
          example: Admin User
        email:
          type: string
          format: email
          maxLength: 254
          description: E-Mail-Adresse des Benutzers.
          example: admin@example.local
        department:
          type: string
          minLength: 1
          maxLength: 200
          description: Abteilung des Benutzers.
          example: Risk Management
        role:
          type: string
          enum:
            - Admin
            - User
          description: Rolle des Benutzers. Nur Admins dürfen Einstellungen und Benutzer
            verwalten.
          example: Admin
        active:
          type: boolean
          description: Steuert, ob sich der Benutzer anmelden darf.
          example: true
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: Passwort des Benutzers. Wird nie in API-Antworten zurückgegeben.
          example: admin123
      example:
        active: false
    UpdateProfile:
      title: UpdateProfile
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 80
          pattern: ^[a-zA-Z0-9._-]+$
          description: Eindeutiger Benutzername für die Anmeldung.
          example: admin
        name:
          type: string
          minLength: 1
          maxLength: 200
          description: Anzeigename des Benutzers.
          example: Admin User
        email:
          type: string
          format: email
          maxLength: 254
          description: E-Mail-Adresse des Benutzers.
          example: admin@example.local
        department:
          type: string
          minLength: 1
          maxLength: 200
          description: Abteilung des Benutzers.
          example: Risk Management
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: Passwort des Benutzers. Wird nie in API-Antworten zurückgegeben.
          example: admin123
      example:
        name: Lena Vogt
        department: Risk Office
    LoginRequest:
      title: LoginRequest
      type: object
      additionalProperties: false
      properties:
        username:
          type: string
          minLength: 1
          description: Benutzername oder E-Mail-Adresse.
          example: admin
        password:
          type: string
          minLength: 8
          maxLength: 200
          description: Passwort des Benutzers. Wird nie in API-Antworten zurückgegeben.
          example: admin123
      required:
        - username
        - password
    LoginResponse:
      title: LoginResponse
      type: object
      additionalProperties: false
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          minLength: 1
          description: Bearer-Token für nachfolgende API-Anfragen.
          example: eyJzdWIiOiIxIn0.signature
      required:
        - user
        - token
...
paths:
...
  /api/auth/login:
    post:
      operationId: loginUser
      summary: Benutzer anmelden
      tags:
        - Benutzer
      description: Prüft Benutzername/E-Mail und Passwort. Deaktivierte Benutzer
        können sich nicht anmelden.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/profile:
    get:
      operationId: getProfile
      summary: Eigenes Profil abrufen
      tags:
        - Benutzer
      description: Liefert das Profil des angemeldeten Benutzers ohne Passwortdaten.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Benutzer gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      summary: Eigenes Profil bearbeiten
      tags:
        - Benutzer
      description: Aktualisiert Profilfelder oder Passwort des angemeldeten Benutzers.
        Rolle und Aktivstatus bleiben Admins vorbehalten.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfile"
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Benutzer gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users:
    get:
      operationId: listUsers
      summary: Benutzer auflisten
      tags:
        - Benutzer
      description: Liefert alle Benutzer ohne Passwortdaten.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Nur Admins dürfen Benutzer verwalten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createUser
      summary: Benutzer anlegen
      tags:
        - Benutzer
      description: Legt einen Benutzer mit gehashtem Passwort an.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Nur Admins dürfen Benutzer verwalten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/users/{id}:
    put:
      operationId: updateUser
      summary: Benutzer bearbeiten
      tags:
        - Benutzer
      description: Aktualisiert Benutzerstammdaten, Passwort oder Aktivstatus.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUser"
      parameters:
        - schema:
            type: string
            minLength: 1
          in: path
          name: id
          required: true
          description: Technische ID des Benutzers.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Die Anfrage enthält ungültige oder unvollständige Daten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Die Anmeldedaten sind ungültig oder der Benutzer ist deaktiviert.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Nur Admins dürfen Benutzer verwalten.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Unter der angegebenen ID wurde kein Benutzer gefunden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unerwarteter interner Serverfehler.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
servers:
  - url: http://127.0.0.1:8000
    description: Lokale Entwicklungsumgebung
...